# backend/spring/final-project/src/main/java/com/iot/doc/openapi.yml
openapi: 3.0.3
info:
  title: IoT Mate Device API
  description: >
    Backend API for the IoT Mate Device (ESP32-based).
    POST endpoints are consumed by the ESP32 firmware.
    GET endpoints are consumed by the frontend dashboard.
  version: 1.0.0
paths:
...
  /sessions:
    post:
      summary: Open a new mate session (SYSTEM_STARTED)
      description: >
        Called by the ESP32 when the system starts. Fails if a SYSTEM_STARTED session
        already exists (only one active session is allowed at a time).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MateSessionRequest'
      responses:
        '201':
          description: Session created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MateSessionResponse'
        '400':
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'

    get:
      summary: Get all mate sessions ordered by most recent first
      responses:
        '200':
          description: List of sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MateSessionResponse'
  /sessions/finish:
    post:
      summary: Close the active mate session (SYSTEM_STOPPED)
      description: >
        Called by the ESP32 when the system stops. Updates totalPours on the active
        session, marks it as SYSTEM_STOPPED, and triggers metric computation
        (average temperature, median temperature, anomaly count) via the observer pattern.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MateSessionRequest'
      responses:
        '200':
          description: Session closed and metrics computed
        '400':
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
  /sessions/latest:
    get:
      summary: Get the most recent mate session
      responses:
        '200':
          description: Latest session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MateSessionResponse'
        '404':
          description: No sessions found

  # ===== METRICS =====
...
components:
  schemas:
...
    MateSessionRequest:
      type: object
      required:
        - sessionType
      properties:
        sessionType:
          $ref: '#/components/schemas/SessionType'
        totalPours:
          type: integer
          nullable: true
          description: Total number of pour button presses during the session (required on finish)
    MateSessionResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        sessionType:
          $ref: '#/components/schemas/SessionType'
        totalPours:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
    SessionType:
      type: string
      enum:
        - SYSTEM_STARTED
        - SYSTEM_STOPPED

    # ===== METRIC SCHEMAS =====
...
    ValidationError:
      type: object
      description: Map of field name → error message returned on 400 responses
      additionalProperties:
        type: string
      example:
        temperature: "must be less than or equal to 100.0"
        sessionType: "must not be null"